Add four currently failing FeatureStore tests.
These tests are modelled after the contents of the `grid/feature-grid.html`
example, which shows wrong syncing between layer features and store records
if one either adds manually a record to the store or a feature to the layer
with code like this issued in the console:

    mapPanel.map.layers[1].addFeatures([
        new OpenLayers.Feature.Vector(
            new OpenLayers.Geometry.Point(5,45)
        )
    ]);

The layer would afterwards only have one feature, but the store would still
have the initial 10 records.

Adding a record manually (e.g. with `store.add()`) will also show the wrong
behaviour.