Enable headless testing through casperjs.
This enables us to test 1010 of a total of 1062 single tests without opening the
test suite in a browser. We skip some tests as we currently are not able to mock-up
certain calls of the ancient Test.AnotherWay suite. The total running time is usually
reduced by about 50%.

Call the script from the root of a GeoExt checkout as follows:

  ./tests/headless/casperjs-1.1-beta.1/bin/casperjs test tests/headless/run-testsuite.js

This also marks the first important step towards continous integration tests, e.g. through
Travis.