Change inclusion method of ExtJS libs&resources
Replaced the old method of inlcuding local files with a copy of the
include-ext.js from the ext-examples. It will now load sources from
http://cdn.sencha.io.

A little slower, but I think it's an enhancement, because a local Ext-Version
will no longer be needed.

The examples can now be loaded with the url param "extjs" to run either
with 4.2.1 or 5.0.1 (maybe some other versions will work to). Default
version is 5.0.1 if no param is passed.

For example:

foo/geoext2/examples/zoomslider/zoomslider.html?extjs=4.2.1

Some themes are only available in ExtJS4 (acces) or ExtJS5 (crisp/
crisp-touch/neptune-touch). So i made the options-toolbar (theme chooser)
react on the param, too.