Changed inclusion method of ExtJS libs&resources
Replaced the old method of inlcuding local files with a copy of the include-ext.js
from the ext-examples. It will now load sources from http://cdn.sencha.io.
A little slower, but I think it's an enhancement, cause a local Ext-Version
would no longer be needed.

thx for the tip @bjornharrtell