Rename bind and unbind functions in Slider classes
This renames the 'bind' and 'unbind' functions within the slider classes
of GeoExt because in ExtJS 5 it seems that 'bind' and 'unbind' are a kind
of reserved config properties (see
http://docs.sencha.com/extjs/5.0.1/#!/api/Ext.slider.Single-cfg-bind) so
our overwrite with a function enforces an exception in Ext.mixins.Bindable when
destroying a slider with its 'destroy' function.
Since 'bind' and 'unbind' were private functions a renaming seems
appropriate.