Optimization of inclusion method for ExtJS libs
Used the changes from ca80980049e625021e7365cb69646fa567b83eb2 and
made some further enhancments.

The examples can now be loaded with the url param "extjs" to run either
with 4.2.1 or 5.0.1 (maybe some other versions will work to). Default
version is 5.0.1 if no param is passed.

For example:

foo/geoext2/examples/zoomslider/zoomslider.html?extjs=4.2.1

Some themes are only available in ExtJS4 (acces) or ExtJS5 (crisp/
crisp-touch/neptune-touch). So i made the options-toolbar (theme chooser)
react on the param, too.

The include-ext.js also works with the tests and so i'm going to change
the inclusion-method for these in an upcoming commit.

Some minor changes to code-indentation or removale of dangling
whitespaces are also inculded.