Fix event forwarding for layername changes.
Previously we wouldn't be notified when the name of a layer changed,
as we erroneously compared `e.property` with 'nathis' instead of
'name' inside of `onChangelayer`. I think the original cause of this
was a search and replace gone wrong in e2712c0ba605a9d35681d9560e3e.

This commit also adds a test for the event forwarding, so we don't
get bitten again.