Using the new lib-inclusion method for the tests
Basically this commit includes codechanges to get the tests running with the
inclusion method from ca80980049e625021e7365cb69646fa567b83eb2.
This includes changes in all the tests and the setup of the testsuite in
tests/Test.AnotherWay.js.

Additonally some sourcecode is modified to get more tests runnning with
ExtJS5.0.1:

- this.model = 'foo.bar' is replace with this.setModel('foo.bar')
- bind an unbind functions in selection/FeatureModel.js are renamed according
  to 3887196b47126440f3256fd26a182dd5232b337d
- some commits are work in progress but maybe someone can benefit from these
  tests failing with ExtJS5.0.1:
  - Form.html
  - data/proxy/Protocol.hmtl
  - data/reader/Attribute.html
  - data/reader/WfsCapabilities.html
  - data/reader/Wmc.html
  - data/reader/WmsCapabilities.html
  - data/reader/WmsDescribeLayer.html
  - form/action/Search.html
  - form/Basic.html
  - panel/map.html
  - selection/FeatureModel.html
  - tree/LayerContainer.html