Init MVVM example app with ExtJS5
This introduces the base setup for an ExtJS 5 based example app.
The project structure is created with Sencha Cmd and the containing files
are only the raw code files. No compiled sources are included here.
The package dependency to GeoExt is also missing here to avoid redundancy.
This will be solved by a later commit or the steps to resolve the GeoExt
package dependency manually will be described in a README file.